Removing the local SoapControllerBase as it was just intended for testing/learning.  Added constants for the camel routes.
