+++ src/main/java/gov/cdc/izgateway/xform/camel/constants/Components.java
@@ -1,0 +1,7 @@
+package gov.cdc.izgateway.xform.camel.constants;
+
+public class Components {
+    public static final String DIRECT = "direct";
+    public static final String IZGHUB = "izghub";
+    public static final String IIS = "iis";
+}
