+++ src/main/java/gov/cdc/izgateway/xform/camel/constants/EndpointPaths.java
@@ -1,0 +1,8 @@
+package gov.cdc.izgateway.xform.camel.constants;
+
+public class EndpointPaths {
+    public static final String IZGHUB_TRANSFORMER_PIPELINE = "izghubTransformerPipeline";
+    public static final String IIS_TRANSFORMER_PIPELINE = "iisTransformerPipeline";
+    public static final String HUB_SERVICE = "IISHubService";
+    public static final String IIS_SERVICE = "IISService";
+}
